# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Base Controller",
        "description": "Controller for basic functionality, such as validation and authorization.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
        "/api/campaigns": {
            "get": {
                "tags": [
                    "Campaigns"
                ],
                "summary": "List all campaigns (paginated)",
                "operationId": "abf2552180345bbac2be4420cde9aad2",
                "responses": {
                    "200": {
                        "description": "Campaign list retrieved"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Campaigns"
                ],
                "summary": "Create a new campaign",
                "operationId": "7658d401faad3d17f624722c1d17e899",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "amount",
                                    "contract_fee",
                                    "interest_rate",
                                    "campaign_type",
                                    "deadline",
                                    "loan_duration"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "My Campaign"
                                    },
                                    "amount": {
                                        "type": "integer",
                                        "example": 1000
                                    },
                                    "contract_fee": {
                                        "type": "integer",
                                        "example": 100
                                    },
                                    "interest_rate": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 5.5
                                    },
                                    "campaign_type": {
                                        "type": "string",
                                        "enum": [
                                            "normal",
                                            "auction"
                                        ]
                                    },
                                    "deadline": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-06-30"
                                    },
                                    "loan_duration": {
                                        "type": "integer",
                                        "example": 12
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Campaign created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/campaigns/{id}": {
            "get": {
                "tags": [
                    "Campaigns"
                ],
                "summary": "Get campaign details",
                "operationId": "feb49ed63db04e9b18ab01a36b2750e4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Campaign ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Campaign found"
                    },
                    "404": {
                        "description": "Campaign not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/campaigns/{id}/close": {
            "post": {
                "tags": [
                    "Campaigns"
                ],
                "summary": "Close a campaign",
                "operationId": "321ccbcff0518b671d93282b4277288b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Campaign ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Campaign closed or cancelled"
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/campaigns/{id}/cancel": {
            "post": {
                "tags": [
                    "Campaigns"
                ],
                "summary": "Cancel a campaign",
                "operationId": "e413db69419740dd427c16eb323f0b96",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Campaign ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Campaign cancelled"
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/campaigns/{campaignId}/finalize": {
            "post": {
                "tags": [
                    "Campaigns"
                ],
                "summary": "Finalize a campaign and create a loan",
                "operationId": "413c1d60e9b1b5f1db94297ecdff6eb1",
                "parameters": [
                    {
                        "name": "campaignId",
                        "in": "path",
                        "description": "Campaign ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Campaign finalized and loan created"
                    },
                    "400": {
                        "description": "Campaign not fully funded"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "passport": {
                "type": "oauth2",
                "description": "Laravel Passport OAuth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost/oauth/authorize",
                        "tokenUrl": "http://localhost/oauth/token",
                        "refreshUrl": "http://localhost/token/refresh",
                        "scopes": []
                    }
                }
            },
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "passport": []
        }
    ]
}